# src/core/http/api/challenges/mission/schemas/mission_community.py
from datetime import datetime

from pydantic import BaseModel, Field

from src.core.enum.challenges.status import OccupancyStatusEnum


class MissionCommunityCreateObject(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    MissionCommunityCreateDTO - a model defined in OpenAPI

        mission_id: The mission_id of this MissionCommunityCreateDTO .
        author: The author of this MissionCommunityCreateDTO .
        place: The place of this MissionCommunityCreateDTO .
        meeting_date: The meeting_date of this MissionCommunityCreateDTO .
        people_required: The people_required of this MissionCommunityCreateDTO .
        people_max: The people_max of this MissionCommunityCreateDTO .
        comment: The comment of this MissionCommunityCreateDTO .
        status: The status of this MissionCommunityCreateDTO .
    """

    mission_id: int = Field(alias="mission_id", default=None)
    author: str = Field(alias="author", default=None)
    place: str = Field(alias="place", default=None)
    meeting_date: datetime = Field(alias="meeting_date", default=None)
    people_required: int = Field(alias="people_required", default=None)
    people_max: int = Field(alias="people_max", default=None)
    comment: str = Field(alias="comment", default=None)
    status: OccupancyStatusEnum = Field(alias="status", default=OccupancyStatusEnum.ACTIVE)


class MissionCommunityEntity(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    MissionCommunityEntity - a model defined in OpenAPI

        id: The id of this MissionCommunityEntity .
        community_id: The community_id of this MissionCommunityEntity .
        mission_id: The mission_id of this MissionCommunityEntity .
        author: The author of this MissionCommunityEntity .
        date_start: The date_start of this MissionCommunityEntity .
        date_close: The date_close of this MissionCommunityEntity .
        status: The status of this MissionCommunityEntity .
        place: The place of this MissionCommunityEntity .
        meeting_date: The meeting_date of this MissionCommunityEntity .
        people_required: The people_required of this MissionCommunityEntity .
        people_max: The people_max of this MissionCommunityEntity .
        comment: The comment of this MissionCommunityEntity .
    """

    id: int = Field(alias="id", default=None)
    community_id: int = Field(alias="community_id", default=None)
    mission_id: int = Field(alias="mission_id", default=None)
    author: str = Field(alias="author", default=None)
    date_start: datetime = Field(alias="date_start", default=None)
    date_close: datetime = Field(alias="date_close", default=None)
    status: OccupancyStatusEnum = Field(alias="status", default=OccupancyStatusEnum.ACTIVE)
    place: str = Field(alias="place", default=None)
    meeting_date: datetime = Field(alias="meeting_date", default=None)
    people_required: int = Field(alias="people_required", default=None)
    people_max: int = Field(alias="people_max", default=None)
    comment: str = Field(alias="comment", default=None)


class MissionCommunityFilterObject(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    MissionCommunityFilter - a model defined in OpenAPI

        community_id: The community_id of this MissionCommunityFilter .
        community_id_list: The community_id_list of this MissionCommunityFilter .
        mission_id: The mission_id of this MissionCommunityFilter .
        status: The status of this MissionCommunityFilter .
    """

    community_id: int = Field(alias="community_id", default=None)
    community_id_list: list[int] = Field(alias="community_id_list", default=None)
    mission_id: int = Field(alias="mission_id", default=None)
    status: OccupancyStatusEnum = Field(alias="status", default=OccupancyStatusEnum.ACTIVE)


class MissionCommunityUpdateObject(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    MissionCommunityUpdateDTO - a model defined in OpenAPI

        date_close: The date_close of this MissionCommunityUpdateDTO .
        status: The status of this MissionCommunityUpdateDTO .
        place: The place of this MissionCommunityUpdateDTO .
        meeting_date: The meeting_date of this MissionCommunityUpdateDTO .
        people_required: The people_required of this MissionCommunityUpdateDTO .
        people_max: The people_max of this MissionCommunityUpdateDTO .
        comment: The comment of this MissionCommunityUpdateDTO .
    """

    date_close: datetime = Field(alias="date_close", default=None)
    status: OccupancyStatusEnum = Field(alias="status", default=OccupancyStatusEnum.ACTIVE)
    place: str = Field(alias="place", default=None)
    meeting_date: datetime = Field(alias="meeting_date", default=None)
    people_required: int = Field(alias="people_required", default=None)
    people_max: int = Field(alias="people_max", default=None)
    comment: str = Field(alias="comment", default=None)