# src/core/http/api/challenges/mission/schemas/mission_user.py
from datetime import datetime

from pydantic import BaseModel, Field

from src.core.enum.challenges.status import OccupancyStatusEnum


class MissionUserCreateObject(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    MissionUserCreateDTO - a model defined in OpenAPI

        mission_id: The mission_id of this MissionUserCreateDTO [Optional.
        status: The status of this MissionUserCreateDTO [Optional.
    """

    mission_id: int = Field(alias="mission_id", default=None)
    status: OccupancyStatusEnum = Field(alias="status", default=OccupancyStatusEnum.ACTIVE)


class MissionUserEntity(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    MissionUserEntity - a model defined in OpenAPI

        id: The id of this MissionUserEntity [Optional.
        user_id: The user_id of this MissionUserEntity [Optional.
        mission_id: The mission_id of this MissionUserEntity [Optional.
        date_start: The date_start of this MissionUserEntity [Optional.
        date_close: The date_close of this MissionUserEntity [Optional.
        status: The status of this MissionUserEntity [Optional.
    """

    id: int = Field(alias="id", default=None)
    user_id: int = Field(alias="user_id", default=None)
    mission_id: int = Field(alias="mission_id", default=None)
    date_start: datetime = Field(alias="date_start", default=None)
    date_close: datetime = Field(alias="date_close", default=None)
    status: OccupancyStatusEnum = Field(alias="status", default=OccupancyStatusEnum.ACTIVE)


class MissionUserFilterObject(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    MissionUserFilter - a model defined in OpenAPI

        mission_id: The mission_id of this MissionUserFilter [Optional.
        status: The status of this MissionUserFilter [Optional.
    """

    mission_id: int = Field(alias="mission_id", default=None)
    status: OccupancyStatusEnum = Field(alias="status", default=OccupancyStatusEnum.ACTIVE)


class MissionUserUpdateObject(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    MissionUserUpdateDTO - a model defined in OpenAPI

        date_close: The date_close of this MissionUserUpdateDTO [Optional.
        status: The status of this MissionUserUpdateDTO [Optional.
    """

    date_close: datetime = Field(alias="date_close", default=None)
    status: OccupancyStatusEnum = Field(alias="status", default=OccupancyStatusEnum.ACTIVE)